Add setCipher method.
The API requires that all client SSL connections use the rsa_rc4_128_sha
cipher. This should not be an issue as curl will use this by default. In those
situations where this is not the case the setCipher method can be called
to force curl to use the correct cipher.